requirements.txt

fastapi==0.95.2
uvicorn==0.23.0
langchain-openai
crewai-tools
python-dotenv
